refactor(item.details): refactor item details page
Linked to #33
