+++ public/app.js
@@ -1,11 +1,12 @@
 import angular from 'angular';
 import SearchModule from './search/search';
 import PodcastsModule from './podcasts/podcasts';
+import ItemModule from './item/item';
 
 let app = angular.module('podcastApp', [
     SearchModule.name,
     PodcastsModule.name,
-    'ps.item',
+    ItemModule.name,
     'ps.download',
     'ps.player',
     'ps.stats',
