+++ public/item/details/item.details.js
@@ -1,78 +1,15 @@
-class ItemDetailCtrl {
 
-    constructor($scope, DonwloadManager, $location, playlistService, podcast, item){
-        this.item = item;
-        this.$location = $location;
-        this.item.podcast = podcast;
-        this.playlistService = playlistService;
-        this.DonwloadManager = DonwloadManager;
+import angular from 'angular';
+import DownloadManager from 'common/service/data/downloadManager';
+import PlaylistService from 'common/service/playlistService';
+import HtmlFilters from 'common/filter/html2plainText';
+import ItemDetailCtrl from './item.details.controller';
 
-        //** WebSocket Inscription **//
-        let webSockedUrl = "/topic/podcast/".concat(this.item.podcast.id);
-
-        this.DonwloadManager
-            .ws
-            .subscribe(webSockedUrl, (message) => {
-                let itemFromWS = JSON.parse(message.body);
-                if (itemFromWS.id == this.item.id) {
-                    _.assign(this.item, itemFromWS);
-                }
-            }, $scope);
-
-    }
-
-    stopDownload(item) {
-        this.DonwloadManager.ws.stop(item);
-    }
-
-    toggleDownload(item) {
-        this.DonwloadManager.ws.toggle(item);
-    }
-
-    remove(item) {
-        return item.remove()
-            .then(() => {
-            this.playlistService.remove(item);
-            this.$location.path('/podcasts/'.concat(this.item.podcast.id));
-        });
-    }
-
-    reset(item) {
-        return item.reset()
-            .then((itemReseted) => {
-                _.assign(this.item, itemReseted);
-                this.playlistService.remove(item);
-            });
-    }
-
-    toggleInPlaylist() {
-        this.playlistService.addOrRemove(this.item);
-    }
-
-    isInPlaylist() {
-        return this.playlistService.contains(this.item);
-    }
-}
-
-angular.module('ps.item.details', [
-    'ps.common.service.data.downloadManager',
-    'ps.common.service.playlist',
-    'ps.common.filter.html2plainText'
-]).config(function($routeProvider, commonKey) {
-    $routeProvider.
-        when('/podcasts/:podcastId/item/:itemId', {
-            templateUrl: 'item/details/item-details.html',
-            controller: 'ItemDetailCtrl',
-            controllerAs: 'idc',
-            hotkeys: commonKey,
-            resolve : {
-                item : function (itemService, $route) {
-                    return itemService.findById($route.current.params.podcastId, $route.current.params.itemId);
-                },
-                podcast : function (podcastService, $route) {
-                    return podcastService.findById($route.current.params.podcastId);
-                }
-            }
-        });
-})
+export default angular
+    .module('ps.item.details', [
+        DownloadManager.name,
+        PlaylistService.name,
+        HtmlFilters.name
+    ])
+    .config(ItemDetailCtrl.routeConfig)
     .controller('ItemDetailCtrl', ItemDetailCtrl);
