+++ public/item/item.js
@@ -2,7 +2,9 @@
 *  Created by kevin on 01/11/14 for Podcast Server
 */
 
-angular.module('ps.item', [
+import ItemDetailsModule from './details/item.details';
+
+export default angular.module('ps.item', [
     'ps.item.details',
     'ps.item.player'
 ]);
